Move signMutilsig view
