+++ src/modules/wallet/components/signMultisigForm/styles.css
@@ -1,5 +1,5 @@
-@import '../styles.css';
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../signMultisigView/styles.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .input {
   display: none;
@@ -28,7 +28,7 @@
     cursor: pointer;
     width: 100%;
     height: 100%;
-    background: url(../../../../../../setup/react/assets/images/icons/upload.svg) center no-repeat;
+    background: url(../../../../../setup/react/assets/images/icons/upload.svg) center no-repeat;
     resize: none;
     border: none;
     box-sizing: border-box;
