+++ src/modules/wallet/components/signMultisigForm/testRegisterMultisgn.json
