+++ src/modules/wallet/components/signMultisigStatus/styles.css
@@ -1,4 +1,4 @@
-@import '../styles.css';
+@import '../signMultisigView/styles.css';
 
 .content {
   margin: 0 auto;
