+++ src/modules/wallet/components/signMultisigSummary/index.js
