+++ src/modules/wallet/components/signMultisigSummary/styles.css
