+++ src/modules/wallet/components/signMultisigSummary/summary.test.js
