+++ src/modules/wallet/components/signMultisigView/helpers.js
