+++ src/modules/wallet/components/signMultisigView/progressBar.js
