+++ src/modules/wallet/components/signMultisigView/styles.css
@@ -1,0 +1,59 @@
+.progressBarContainer {
+  margin: 5px 0 15px 0;
+}
+
+.boxContainer {
+  & > header {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 20px 20px 0 20px;
+    height: auto;
+    overflow-y: auto;
+
+    & > p {
+      margin-bottom: 0;
+      color: var(--color-maastricht-blue);
+    }
+  }
+
+  & > .footer {
+    padding-top: 20px;
+
+    & + footer {
+      padding-top: 0;
+    }
+  }
+
+  & .feedback {
+    width: 100%;
+    padding: 0 var(--horizontal-padding-m);
+    box-sizing: border-box;
+    margin-bottom: 10px;
+
+    & > span {
+      width: 100%;
+      color: var(--color-strong-white);
+      border-radius: var(--border-radius-standard);
+      display: block;
+      padding: var(--vertical-padding-s) var(--horizontal-padding-m);
+      box-sizing: border-box;
+    }
+
+    &.error > span {
+      background: var(--error-background-color);
+    }
+
+    &.warning > span {
+      background: var(--color-danger);
+    }
+  }
+}
+
+.txDetails {
+  max-height: calc(100vh - 80px - 290px); /* stylelint-disable-line unit-whitelist */
+
+  &.small {
+    max-height: calc(100vh - 110px - 290px); /* stylelint-disable-line unit-whitelist */
+  }
+}
