+++ src/modules/wallet/manager/signMultisigManager.js
@@ -1,0 +1,4 @@
+import { withRouter } from 'react-router-dom';
+import SignMultisigView from '../components/signMultisigView';
+
+export default withRouter(SignMultisigView);
